split UI framework into 2D and 3D
