+++ packages/ui.js
@@ -12,6 +12,6 @@ import { Deck } from '../src/ui/deck.js';
 import { Grid } from '../src/ui/grid.js';
 import { HexGrid } from '../src/ui/hex.js';
 import { HexUtils } from '../src/ui/hex-utils.js';
-import Token from '../src/ui/token.js';
+import { Token } from '../src/ui/token.js';
 
 export { UI, Card, Deck, Grid, HexGrid, Token, HexUtils };
