+++ src/ui/2d/grid.test.js
@@ -8,7 +8,7 @@
 
 import React from 'react';
 import { Grid } from './grid';
-import Token from './token';
+import { Token } from './token';
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
