+++ src/ui/2d/hex.test.js
@@ -8,7 +8,7 @@
 
 import React from 'react';
 import { HexGrid } from './hex';
-import Token from './token';
+import { Token } from './token';
 import Enzyme from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
