+++ src/ui/3d/card.css
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2017 The boardgame.io Authors
+ *
+ * Use of this source code is governed by a MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+.bgio-card {
+  display: flex;
+  user-select: none;
+  font-family: monospace;
+  font-weight: bold;
+  font-size: 18px;
+  color: #ababab;
+  text-align: center;
+  flex-direction: column;
+  justify-content: center;
+  cursor: pointer;
+  background: #fff;
+  border-radius: 6px;
+  border: 1px solid #cdcdcd;
+  width: 100px;
+  height: 140px;
+  overflow: hidden;
+  transition: transform 0.1s;
+}
+
+.bgio-card.placeholder {
+  cursor: default;
+  opacity: 0;
+  pointer-events: none;
+}
+
+.bgio-card.accept {
+  transform: rotate(10deg);
+  box-shadow: 5px 5px 5px #ddd;
+}
+
+.bgio-card.reject {
+}
+
+.bgio-card__front,
+.bgio-card__back {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  overflow: hidden;
+}
+
+.bgio-card__back {
+  background-image: url("data:image/svg+xml,%3Csvg width='20' height='20' viewBox='0 0 64 64' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath d='M8 16c4.418 0 8-3.582 8-8s-3.582-8-8-8-8 3.582-8 8 3.582 8 8 8zm0-2c3.314 0 6-2.686 6-6s-2.686-6-6-6-6 2.686-6 6 2.686 6 6 6zm33.414-6l5.95-5.95L45.95.636 40 6.586 34.05.636 32.636 2.05 38.586 8l-5.95 5.95 1.414 1.414L40 9.414l5.95 5.95 1.414-1.414L41.414 8zM40 48c4.418 0 8-3.582 8-8s-3.582-8-8-8-8 3.582-8 8 3.582 8 8 8zm0-2c3.314 0 6-2.686 6-6s-2.686-6-6-6-6 2.686-6 6 2.686 6 6 6zM9.414 40l5.95-5.95-1.414-1.414L8 38.586l-5.95-5.95L.636 34.05 6.586 40l-5.95 5.95 1.414 1.414L8 41.414l5.95 5.95 1.414-1.414L9.414 40z' fill='%23ababab' fill-opacity='0.4' fill-rule='evenodd'/%3E%3C/svg%3E");
+  background-position: 2px 2px;
+  outline: 8px solid #eee;
+  outline-offset: -20px;
+}
