+++ src/ui/3d/deck.css
@@ -1,0 +1,20 @@
+/*
+ * Copyright 2017 The boardgame.io Authors
+ *
+ * Use of this source code is governed by a MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+.bgio-deck {
+  border: 1px dashed #ddd;
+  position: relative;
+  display: inline-flex;
+  border-radius: 6px;
+  padding: 5px;
+  margin-top: 20px;
+  margin-bottom: 20px;
+  margin-right: 20px;
+  width: 100px;
+  height: 140px;
+}
