+++ src/ui/3d/hex.js
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2018 The boardgame.io Authors
+ *
+ * Use of this source code is governed by a MIT-style
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+import React from 'react';
+
+// Not yet implemented.
+export class HexGrid extends React.Component {
+  render() {
+    return null;
+  }
+}
+
+// Not yet implemented.
+export class Hex extends React.Component {
+  render() {
+    return null;
+  }
+}
