+++ src/ui/3d/token.js
@@ -1,0 +1,16 @@
+/*
+ * Copyright 2018 The boardgame.io Authors
+ *
+ * Use of this source code is governed by a MIT-syle
+ * license that can be found in the LICENSE file or at
+ * https://opensource.org/licenses/MIT.
+ */
+
+import React from 'react';
+
+// Not yet implemented.
+export class Token extends React.Component {
+  render() {
+    return null;
+  }
+}
